Refactor generator
